backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.6.1
pymysql==1.1.0
cryptography==42.0.5
azure-storage-blob==12.19.0
python-dotenv==1.0.1